refactor(pipeline): centralize construction-time stacks in a pipeline module
Splits pipeline.rs into a directory module and moves the
assemble-phase stage builders in beside the phases, so every
construction-time tower stack reads from one module: mod.rs holds
build_pipeline, acquire.rs the fallible acquisition helpers, and
stages.rs the stage builders including build_supergraph_pipeline,
build_execution_service, and build_supergraph_service (from
services/supergraph/service.rs) and query_parsing_service (from
services/query_parsing.rs, which keeps its types).

The router stacks stay with RouterCreator (the persistent struct owns
its constructor), and the subgraph and HTTP client folds stay in
their factories: the subgraph fold lives inside the factory
FetchService consumes, and the HTTP client fold runs lazily per
connector request, so neither is construction-time assembly.
