+++ apollo-router/src/services/supergraph/service.rs
@@ -13,15 +13,12 @@ use indexmap::IndexMap;
 use opentelemetry::Key;
 use opentelemetry::KeyValue;
 use tower::BoxError;
-use tower::ServiceBuilder;
 use tower::ServiceExt;
 use tower::load_shed::error::Overloaded;
 use tower_service::Service;
 use tracing_futures::Instrument;
 
-use crate::Configuration;
 use crate::Context;
-use crate::batching::BatchQueryPlanAnalysisLayer;
 use crate::compute_job::ComputeBackPressureError;
 use crate::configuration::mode::Mode;
 use crate::error::CacheResolverError;
@@ -29,43 +26,21 @@ use crate::graphql;
 use crate::graphql::IntoGraphQLErrors;
 use crate::introspection;
 use crate::introspection::IntrospectionService;
-use crate::layers::InternalServiceBuilderExt as _;
-use crate::layers::ServiceBuilderExt as _;
-use crate::layers::unconstrained_buffer::UnconstrainedBuffer;
 use crate::plugin::DynPlugin;
 use crate::plugins::connectors::query_plans::store_connectors;
 use crate::plugins::connectors::query_plans::store_connectors_labels;
-use crate::plugins::limits::operation_limits_layer::EnforceOperationLimitsLayer;
-use crate::plugins::subscription::APOLLO_SUBSCRIPTION_PLUGIN;
-use crate::plugins::subscription::Subscription;
-use crate::plugins::subscription::SubscriptionExecutionLayer;
-use crate::plugins::telemetry::Telemetry;
 use crate::plugins::telemetry::config_new::events::log_event;
 use crate::plugins::telemetry::config_new::supergraph::events::SupergraphEventResponse;
 use crate::plugins::telemetry::consts::QUERY_PLANNING_SPAN_NAME;
-use crate::query_planner::CachingQueryPlanner;
-use crate::query_planner::InMemoryQueryPlanCache;
-use crate::query_planner::QueryPlanCache;
-use crate::query_planner::SubgraphSchemas;
 use crate::services::ExecutionRequest;
 use crate::services::ExecutionResponse;
 use crate::services::QueryPlannerResponse;
-use crate::services::SubgraphServiceFactory;
 use crate::services::SupergraphRequest;
 use crate::services::SupergraphResponse;
-use crate::services::connector::request_service::ConnectorRequestServiceFactory;
-use crate::services::connector_service::ConnectorServiceFactory;
 use crate::services::execution;
-use crate::services::execution::service::ExecutionService;
-use crate::services::fetch_service::FetchService;
-use crate::services::http::HttpClientServiceFactory;
-use crate::services::layers::allow_only_http_post_mutations::AllowOnlyHttpPostMutationsLayer;
-use crate::services::layers::content_negotiation;
 use crate::services::query_parsing::ParsedDocument;
 use crate::services::query_planner;
 use crate::services::router::ClientRequestAccepts;
-use crate::services::subgraph;
-use crate::services::supergraph;
 use crate::spec::Schema;
 
 pub(crate) const FIRST_EVENT_CONTEXT_KEY: &str = "apollo::supergraph::first_event";
@@ -446,176 +421,4 @@ async fn plan_query(
         .await?;
 
     Ok(qpr)
-}
-
-/// Assembles the supergraph pipeline: wraps the query planner in `query_plan_cache`, then
-/// builds the execution and supergraph service stacks.
-///
-/// Part of the assemble phase in [`crate::pipeline`]; call it after plugin activation, with
-/// a cache built by [`crate::pipeline::build_query_plan_cache`].
-///
-/// Returns three values:
-///
-/// - the buffered supergraph service; clone it for each consumer
-/// - a handle on the in-memory query-plan cache, for warm-up on the next reload
-/// - the caching query planner service, for query-plan cache warm-up
-#[allow(clippy::too_many_arguments)]
-pub(crate) fn build_supergraph_pipeline(
-    query_planner_service: query_planner::BoxCloneService,
-    query_plan_cache: QueryPlanCache,
-    schema: Arc<Schema>,
-    subgraph_schemas: Arc<SubgraphSchemas>,
-    configuration: Arc<Configuration>,
-    plugins: Arc<Plugins>,
-    subgraph_services: Vec<(String, subgraph::BoxCloneService)>,
-    connector_http_service_factory: IndexMap<String, HttpClientServiceFactory>,
-) -> (
-    supergraph::BoxCloneService,
-    InMemoryQueryPlanCache,
-    query_planner::CacheBoxCloneService,
-) {
-    let query_planner_service = CachingQueryPlanner::new(
-        query_planner_service,
-        schema.clone(),
-        subgraph_schemas.clone(),
-        &configuration,
-        query_plan_cache.clone(),
-    )
-    .boxed_clone();
-
-    let introspection_service = introspection::introspection_service(&configuration);
-
-    let execution_service = build_execution_service(
-        schema.clone(),
-        subgraph_schemas,
-        plugins.clone(),
-        subgraph_services,
-        connector_http_service_factory,
-        configuration.clone(),
-    );
-
-    let supergraph_service = build_supergraph_service(
-        query_planner_service.clone(),
-        execution_service,
-        introspection_service,
-        schema,
-        &configuration,
-        plugins,
-    );
-
-    (
-        supergraph_service.boxed_clone(),
-        query_plan_cache.in_memory_cache(),
-        query_planner_service,
-    )
-}
-
-/// Assembles the execution service stack: the batching and subscription layers, each plugin's
-/// `execution_service` hook, and the [`ExecutionService`] with its [`FetchService`] that
-/// dispatches to subgraphs and connectors.
-pub(crate) fn build_execution_service(
-    schema: Arc<Schema>,
-    subgraph_schemas: Arc<SubgraphSchemas>,
-    plugins: Arc<Plugins>,
-    subgraph_services: Vec<(String, subgraph::BoxCloneService)>,
-    connector_http_service_factory: IndexMap<String, HttpClientServiceFactory>,
-    configuration: Arc<Configuration>,
-) -> execution::BoxCloneService {
-    let subscription_plugin_conf = plugins
-        .iter()
-        .find(|i| i.0.as_str() == APOLLO_SUBSCRIPTION_PLUGIN)
-        .and_then(|plugin| (*plugin.1).as_any().downcast_ref::<Subscription>())
-        .map(|p| p.config.clone());
-
-    let fetch_service = FetchService::new(
-        schema.clone(),
-        subgraph_schemas.clone(),
-        Arc::new(SubgraphServiceFactory::new(
-            subgraph_services,
-            plugins.clone(),
-            configuration.notify.clone(),
-            subscription_plugin_conf.clone().map(Arc::new),
-            configuration.apq.subgraph.clone(),
-        )),
-        subscription_plugin_conf.clone(),
-        Arc::new(ConnectorServiceFactory::new(
-            schema.clone(),
-            subgraph_schemas.clone(),
-            subscription_plugin_conf.clone(),
-            schema
-                .connectors
-                .as_ref()
-                .map(|c| c.by_service_name.clone())
-                .unwrap_or_default(),
-            Arc::new(ConnectorRequestServiceFactory::new(
-                Arc::new(connector_http_service_factory),
-                plugins.clone(),
-            )),
-        )),
-        Arc::new(configuration.experimental_hoist_orphan_errors.clone()),
-    );
-
-    let apollo_telemetry_conf = plugins
-        .iter()
-        .find(|i| i.0.as_str() == "apollo.telemetry")
-        .and_then(|plugin| (*plugin.1).as_any().downcast_ref::<Telemetry>())
-        .map(|t| t.config.apollo.clone());
-
-    ServiceBuilder::new()
-        .layer(BatchQueryPlanAnalysisLayer::new())
-        .layer(SubscriptionExecutionLayer::new(
-            configuration.notify.clone(),
-        ))
-        .rust_plugins(plugins.clone(), |plugin, service| {
-            plugin.execution_service(service)
-        })
-        .service(
-            ExecutionService {
-                schema,
-                fetch_service,
-                subscription_config: subscription_plugin_conf,
-                subgraph_schemas,
-                apollo_telemetry_config: apollo_telemetry_conf,
-                configuration,
-            }
-            .boxed_clone(),
-        )
-        .boxed_clone()
-}
-
-/// Assembles the [`SupergraphService`] stack, outermost first: the buffer, the
-/// content-negotiation and compute-job metrics layers, each plugin's `supergraph_service`
-/// hook, and the mutation-restriction and operation-limit layers.
-pub(crate) fn build_supergraph_service(
-    query_planner_service: query_planner::CacheBoxCloneService,
-    execution_service: execution::BoxCloneService,
-    introspection_service: IntrospectionService,
-    schema: Arc<Schema>,
-    configuration: &Configuration,
-    plugins: Arc<Plugins>,
-) -> UnconstrainedBuffer<SupergraphRequest, BoxFuture<'static, Result<SupergraphResponse, BoxError>>>
-{
-    let supergraph_service = SupergraphService::builder()
-        .query_planner_service(query_planner_service)
-        .execution_service(execution_service)
-        .introspection_service(introspection_service)
-        .schema(schema)
-        .strict_variable_validation(configuration.supergraph.strict_variable_validation)
-        .build();
-
-    // The outer buffer provides backpressure for the full supergraph pipeline and is
-    // required for correct LoadShed / ConcurrencyLimit / RateLimit behaviour introduced
-    // by traffic-shaping and other plugins (see ServiceBuilderExt::buffered).
-    ServiceBuilder::new()
-        .buffered()
-        .layer(content_negotiation::SupergraphContentNegotiationLayer::default())
-        .layer(crate::compute_job::ComputeJobMetricsLayer::new())
-        .rust_plugins(plugins, |plugin, service| {
-            plugin.supergraph_service(service)
-        })
-        .layer(AllowOnlyHttpPostMutationsLayer::default())
-        .layer(EnforceOperationLimitsLayer::new(
-            &configuration.limits.router,
-        ))
-        .service(supergraph_service)
 }
